requirements.txt
# Основные зависимости проекта
python-telegram-bot==20.7
python-docx==1.1.0
requests==2.31.0
python-dotenv==1.0.0
qdrant-client==1.15.1
openai==1.12.0
google-genai>=0.8.0
beautifulsoup4==4.12.2
lxml==4.9.3
schedule==1.2.0
httpx==0.26.0
asyncio-throttle==1.0.2
datetime==5.4
pandas==2.1.4
numpy==1.25.2

# Для астрологических расчетов
pyephem==4.1.5
swisseph==2.10.03.2

# Для работы с базами данных
sqlalchemy==2.0.25
alembic==1.13.1

# Утилиты
loguru==0.7.2
pydantic==2.5.3
